pyproject.toml
# SPDX-FileCopyrightText: Copyright (c) 2023-2026, NVIDIA CORPORATION.
# SPDX-License-Identifier: Apache-2.0

[build-system]
requires = ["setuptools>=77.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "rapids-pre-commit-hooks"
version = "1.4.2"
authors = [
    { name = "RAPIDS Development Team", email = "pypi@rapids.ai" }
]
urls = { homepage = "https://github.com/rapidsai/pre-commit-hooks" }
description = "pre-commit hooks for RAPIDS"
readme = { file = "README.md", content-type = "text/markdown" }
license = "Apache-2.0"
license-files = ["LICENSE"]
classifiers = [
    "Programming Language :: Python :: 3",
]
requires-python = ">=3.11"
dependencies = [
    "PyYAML",
    "bashlex",
    "gitpython",
    "levenshtein",
    "packaging",
    "rich",
    "tomlkit",
]

[project.optional-dependencies]
test = [
    "freezegun",
    "pre-commit",
    "pytest",
    "rapids-pre-commit-hooks[alpha-spec]",
]
alpha-spec = [
    "rapids-metadata>=0.4.0,<0.5.0.dev0",
]

[project.scripts]
verify-alpha-spec = "rapids_pre_commit_hooks.alpha_spec:main"
verify-codeowners = "rapids_pre_commit_hooks.codeowners:main"
verify-coderabbit-gitleaks = "rapids_pre_commit_hooks.coderabbit_gitleaks:main"
verify-conda-yes = "rapids_pre_commit_hooks.shell.verify_conda_yes:main"
verify-copyright = "rapids_pre_commit_hooks.copyright:main"
verify-hardcoded-version = "rapids_pre_commit_hooks.hardcoded_version:main"
verify-pyproject-license = "rapids_pre_commit_hooks.pyproject_license:main"

[tool.setuptools]
packages = { "find" = { where = ["src"] } }

[tool.ruff]
line-length = 79

[tool.ruff.lint]
select = [
    # flake8-unused-args
    "ARG",
    # pycodestyle
    "E",
    # pyflakes
    "F",
]

[tool.pytest.ini_options]
testpaths = [
    "tests",
]
pythonpath = [
    "tests/utils",
]

[[tool.mypy.overrides]]
module = "bashlex.*"
ignore_missing_imports = true
